Add LinkVertexRenderer in order to be able to use custom component for vertices
