--- src/Blazor.Diagrams/Components/LinkVertexWidget.razor
@@ -1,10 +1,0 @@
-﻿<circle cx="@Vertex.Position.X.ToInvariantString()"
-        cy="@Vertex.Position.Y.ToInvariantString()"
-        r="5"
-        fill="@ColorToUse"
-        cursor="move"
-        @ondblclick="OnDoubleClick"
-        @onpointerdown="OnPointerDown"
-        @onpointerup="OnPointerUp"
-        @onpointerdown:stopPropagation
-        @onpointerup:stopPropagation/>
